requirements.txt
Django==1.4
django-openid-auth==0.4
mongoengine==0.6.3
psycopg2==2.4.5
pymongo==2.1.1
python-openid==2.2.5
wsgiref==0.1.2
